Move topic-related functions to `topics.py`
