+++ src/poprox_recommender/topics.py
@@ -1,13 +1,43 @@
+from collections import defaultdict
+from uuid import UUID
+
 from torch.nn.functional import cosine_similarity
 from transformers import AutoModel, AutoTokenizer
 
-from poprox_concepts import Article
+from poprox_concepts import Article, ClickHistory
 from poprox_concepts.domain.topics import GENERAL_TOPICS
 
 
 def extract_general_topics(article: Article):
     article_topics = set([mention.entity.name for mention in article.mentions])
     return article_topics.intersection(GENERAL_TOPICS)
+
+
+def find_topic(past_articles: list[Article], article_id: UUID):
+    # each article might correspond to multiple topic
+    for article in past_articles:
+        if article.article_id == article_id:
+            return extract_general_topic(article)
+
+
+def normalized_topic_count(topic_counts: dict[str, int]):
+    total_count = sum(topic_counts.values())
+    normalized_counts = {key: value / total_count for key, value in topic_counts.items()}
+    return normalized_counts
+
+
+def user_topic_preference(past_articles: list[Article], click_history: ClickHistory) -> dict[str, int]:
+    """Topic preference only based on click history"""
+    clicked_articles = click_history.article_ids  # List[UUID]
+
+    topic_count_dict = defaultdict(int)
+
+    for article_id in clicked_articles:
+        clicked_topics = find_topic(past_articles, article_id)
+        for topic in clicked_topics:
+            topic_count_dict[topic] += 1
+
+    return topic_count_dict
 
 
 def classify_news_topic(model, tokenizer, general_topics, topic):
